+++ .dockerignore
@@ -1,0 +1,5 @@
+/.dvc
+/.mypy_cache
+.pytest_cache
+__pycache__
+/data
