+++ Dockerfile
@@ -1,39 +1,16 @@
 FROM nvcr.io/nvidia/deepstream:6.0-triton
 
-RUN SNIPPET="export PROMPT_COMMAND='history -a' && export HISTFILE=/commandhistory/.bash_history" && echo $SNIPPET >> "/root/.bashrc"
+RUN SNIPPET="export PROMPT_COMMAND='history -a' && export HISTFILE=/commandhistory/.bash_history" \
+    && echo $SNIPPET >> "/root/.bashrc"
 
-ENV LANG=C.UTF-8 \
-    LC_ALL=C.UTF-8 \
-    PATH="${PATH}:/root/.poetry/bin"
+RUN apt-key adv --fetch-keys https://developer.download.nvidia.cn/compute/cuda/repos/ubuntu2004/x86_64/3bf863cc.pub
 
-RUN rm /etc/apt/sources.list.d/cuda.list
-RUN rm /etc/apt/sources.list.d/nvidia-ml.list
-RUN apt-key del 7fa2af80
-RUN apt-get update && apt-get install -y --no-install-recommends wget
-RUN wget https://developer.download.nvidia.com/compute/cuda/repos/ubuntu1804/x86_64/cuda-keyring_1.0-1_all.deb
-RUN dpkg -i cuda-keyring_1.0-1_all.deb
-
-RUN apt update -y
-RUN DEBIAN_FRONTEND=noninteractive apt install -y tzdata
+RUN apt-get update -y
 RUN apt install -y \
-    build-essential \
-    curl \
-    git \
-    libcairo2-dev \
-    libgl1-mesa-glx \
-    software-properties-common
-
-# Install Python 3.9
-RUN add-apt-repository ppa:deadsnakes/ppa
-RUN apt install -y python3.9-dev python3.9-venv
-RUN python3.9 -m ensurepip
-RUN ln -s /usr/bin/python3.9 /usr/local/bin/python
-RUN ln -s /usr/local/bin/pip3.9 /usr/local/bin/pip
-RUN pip install --upgrade pip
-
-# Allow root for Jupyter notebooks
-RUN mkdir /root/.jupyter
-RUN echo "c.NotebookApp.allow_root = True" > /root/.jupyter/jupyter_notebook_config.py
+    python3-gi python3-dev python3-gst-1.0 python-gi-dev git python-dev \
+    python3 python3-pip python3.8-dev python3.8-venv cmake g++ build-essential \
+    libglib2.0-dev libglib2.0-dev-bin python-gi-dev libtool m4 autoconf automake \
+    openssh-client
 
 RUN git config --global http.sslverify false
 
@@ -49,15 +26,24 @@ RUN cd deepstream_python_apps && \
 RUN cd deepstream_python_apps/bindings && \
     mkdir build && \
     cd build && \
-    cmake .. -DPYTHON_MAJOR_VERSION=3 -DPYTHON_MINOR_VERSION=9 && \
+    cmake .. -DPYTHON_MAJOR_VERSION=3 -DPYTHON_MINOR_VERSION=8 && \
     make && \
-    pip install pyds-*.whl
+    pip3 install pyds-*.whl
 
+RUN rm /usr/bin/python /usr/local/bin/pip
+RUN ln -s /usr/bin/python3.8 /usr/local/bin/python
+RUN ln -s /usr/local/bin/pip3.8 /usr/local/bin/pip
+
+# Allow root for Jupyter notebooks
+RUN mkdir /root/.jupyter
+RUN echo "c.NotebookApp.allow_root = True" > /root/.jupyter/jupyter_notebook_config.py
+
+# Install Poetry
 RUN curl -sSL https://raw.githubusercontent.com/python-poetry/poetry/master/install-poetry.py | POETRY_HOME=/opt/poetry python && \
     cd /usr/local/bin && \
     ln -s /opt/poetry/bin/poetry && \
     poetry config virtualenvs.create false
-
+  
 COPY pyproject.toml poetry.lock* ./
 
 # Allow installing dev dependencies to run tests
